+++ src/vonage/sms.py
@@ -3,26 +3,8 @@ from datetime import datetime
 from ._internal import _format_date_param
 
 class Sms:
-    #To init Sms class pass a client reference or a key and secret
-    def __init__(
-        self,
-        client=None,
-        key=None,
-        secret=None,
-        signature_secret=None,
-        signature_method=None
-    ):
-        try:
-            self._client = client
-            if self._client is None:
-                self._client = vonage.Client(
-                    key=key,
-                    secret=secret,
-                    signature_secret=signature_secret,
-                    signature_method=signature_method
-                )
-        except Exception as e:
-            print(f'Error: {str(e)}')
+    def __init__(self, client):
+        self._client = client
     
     def send_message(self, params):
         """