+++ tests/conftest.py
@@ -88,12 +88,6 @@ def numbers(client):
     return vonage.Numbers(client)
 
 @pytest.fixture
-def message_search(client):
-    import vonage
-    
-    return vonage.MessageSearch(client)
-
-@pytest.fixture
 def ussd(client):
     import vonage
     
@@ -110,3 +104,9 @@ def messages(client):
     import vonage
 
     return vonage.Messages(client)
+
+@pytest.fixture
+def redact(client):
+    import vonage
+
+    return vonage.Redact(client)